+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Starting from Rust port of the project, all changes will be put into this file.
 
 ### Changes
 - `SJ`: Fix broken serde on renew field at account subcriptions
+- `SJ`: Allow downloading "expired" chapters if you have subscription for it.
 
 ### Build
 - Pin `windows-sys` dependencies (If you use crates.io, this is already pinned in `0.3.0`)